115/use themes from db for oidc routes (#135)
* move config into its own file to clean up main plugin file

* move routes to their own file

* don't pass provider too addROutes function

* throw view object construction into a file

* WIP - hooking everything up to a generic theme service

* don't use electrolyte to load models

* get themplate from db

* move plugin templates to root templates

* loading layout from filesystem when not provided

* user layout from db

* go ahead and move template model to theme dir

* plugin screens are all now themed

* DRY

* refactoring to use views

* refactor profileFormHandler to be async/await

* refactor to use userprofile view

* converting last controller functions to use views

* threw in a bit of cleanup

* make sure error isn't undefined before checking isBoom

* theme the registration page

* theme the profile page

* themed remaining pages
