+++ api/manifest.js
@@ -1,6 +1,7 @@
-var config = require('./config');
-var formatError = require('./src/lib/format-error');
-var fetchKeystore = require('./src/lib/fetch-keystore');
+const config = require('./config');
+const formatError = require('./src/lib/format-error');
+const fetchKeystore = require('./src/lib/fetch-keystore');
+const handlebars = require('handlebars');
 
 var ioc = require('electrolyte');
 ioc.use(ioc.dir('src/lib'));
@@ -11,6 +12,7 @@ module.exports = Promise.all([
   ioc.create('user/user-oidc-service'),
   ioc.create('oidc-adapter/redis'),
   ioc.create('oidc-adapter/sql'),
+  ioc.create('theme/theme-service'),
   fetchKeystore(),
 ])
   .then(values => ({
@@ -18,7 +20,8 @@ module.exports = Promise.all([
     userService: values[1],
     redisOidcAdapter: values[2],
     sqlOidcAdapter: values[3],
-    keystore: values[4],
+    themeService: values[4],
+    keystore: values[5],
   }))
   .then(lib => ({
     server: {
@@ -70,7 +73,17 @@ module.exports = Promise.all([
         plugin: {
           register: './plugins/openid-connect/openid-connect',
           options: {
+            vision: {
+              engines: {
+                hbs: handlebars
+              },
+              path: './templates',
+              layout: true,
+              layoutPath: './templates/layout',
+              layout: 'default',
+            },
             prefix: 'op',
+            getTemplate: lib.themeService.renderThemedTemplate,
             authenticateUser: lib.userService.authenticate,
             findUserById: lib.userService.findByIdForOidc,
             cookieKeys: config('/oidc/cookieKeys'),
