+++ api/src/application/theme/layout-model.js
@@ -1,0 +1,4 @@
+module.exports = (bookshelf) => ({
+  tableName: 'SIP_layout',
+  idAttribute: 'id',
+});
