+++ api/src/application/theme/template-model.js
@@ -1,0 +1,8 @@
+module.exports = (bookshelf) => ({
+  tableName: 'SIP_template',
+  idAttribute: false,
+
+  layout() {
+    return this.belongsTo('layout', 'layout_id');
+  },
+});
