+++ api/src/application/theme/theme-model.js
@@ -1,0 +1,8 @@
+module.exports = bookshelf => ({
+  tableName: 'SIP_theme',
+  idAttribute: 'id',
+
+  templates() {
+    return this.hasMany('template', 'theme_id')
+  }
+});
