+++ api/src/application/user/user-password-reset-token-model.js
@@ -1,7 +1,4 @@
-module.exports = bookshelf => bookshelf.model('user_password_reset_token', {
+module.exports = bookshelf => ({
   tableName: 'SIP_user_password_reset_token',
   idAttribute: 'token',
 });
-
-module.exports['@singleton'] = true;
-module.exports['@require'] = ['bookshelf'];
