+++ api/src/application/user/user-routes.js
@@ -1,5 +1,6 @@
 const Joi = require('joi');
 const Readable = require('stream').Readable;
+const userFormData = require('./user-form-data');
 
 const queryValidation = {
   client_id : Joi.string().required(),
@@ -9,7 +10,7 @@ const queryValidation = {
   nonce : Joi.string().optional(),
 };
 
-module.exports = (service, controller, userFormData) => {
+module.exports = (service, controller) => {
   return [
     {
       method : 'GET',
@@ -135,10 +136,10 @@ module.exports = (service, controller, userFormData) => {
     {
       method : 'GET',
       path : '/user/reset-password',
-      handler : controller.getResetPasswordForm,
+      handler : controller.getResetPasswordForm('Reset Password'),
       config : {
         validate : {
-          failAction : controller.getResetPasswordForm,
+          failAction : controller.getResetPasswordForm('Reset Password'),
           query : Object.assign({
             token: Joi.string().required(),
             client_id: Joi.string().required(),
@@ -151,7 +152,7 @@ module.exports = (service, controller, userFormData) => {
     {
       method : 'POST',
       path : '/user/reset-password',
-      handler : controller.postResetPasswordForm,
+      handler : controller.postResetPasswordForm('Reset Password'),
       config : {
         validate : {
           payload : {
@@ -164,14 +165,14 @@ module.exports = (service, controller, userFormData) => {
             redirect_uri: Joi.string().required(),
             scope: Joi.string().required(),
           }, queryValidation),
-          failAction : controller.getResetPasswordForm,
+          failAction : controller.getResetPasswordForm('Reset Password'),
         }
       },
     },
     {
       method: 'GET',
       path: '/user/accept-invite',
-      handler: controller.getAcceptInviteForm,
+      handler: controller.getResetPasswordForm('Set Password'),
       config: {
         validate: {
           query: {
@@ -186,20 +187,20 @@ module.exports = (service, controller, userFormData) => {
     {
       method: 'POST',
       path: '/user/accept-invite',
-      handler: controller.postAcceptInviteForm,
+      handler: controller.postResetPasswordForm('Set Password'),
       config: {
         validate: {
           payload : {
             password : Joi.string().min(8).required(),
             pass2 : Joi.any().valid(Joi.ref('password')).required(),
-          }, 
+          },
           query: {
             token: Joi.string().guid().required(),
             client_id: Joi.string().required(),
             redirect_uri: Joi.string().required(),
             scope: Joi.string().required(),
           },
-          failAction: controller.getAcceptInviteForm,
+          failAction: controller.getResetPasswordForm('Set Password'),
         },
       },
     },
@@ -222,5 +223,4 @@ module.exports['@singleton'] = true;
 module.exports['@require'] = [
   'user/user-service',
   'user/user-controller',
-  'user/user-form-data',
 ];
