+++ api/src/lib/bookshelf.js
@@ -1,31 +1,15 @@
 const _ = require('lodash');
-
-// This plugin copies methods in model.prototype.relationships to model.prototype on
-// extend, which allows us to easily enumerate all the relations for a given model
-function relationships(bookshelf) {
-  var originalExtend = bookshelf.Model.extend;
-
-  bookshelf.Model.extend = function(protoProps, constructorProps) {
-    if (protoProps.relationships) {
-      var keys = Object.keys(protoProps.relationships);
-      for (var i = 0, len = keys.length; i < len; i++) {
-        protoProps[keys[i]] = protoProps.relationships[keys[i]];
-      }
-    }
-
-    var model = originalExtend.call(this, protoProps, constructorProps);
-
-    model.prototype.relationships = model.prototype.relationships || {};
-    _.defaults(model.prototype.relationships, this.prototype.relationships || {});
-
-    return model;
-  };
-}
+const models = require('../application/models');
 
 module.exports = (knex) => {
   var bookshelf = require('bookshelf')(knex);
   bookshelf.plugin('registry');
-  bookshelf.plugin(relationships);
+
+  for (let i = 0; i < models.length; i++) {
+    const { name, model } = models[i](bookshelf);
+    bookshelf.model(name, model);
+  }
+
   return bookshelf;
 };
