+++ api/templates/end_session.hbs
